HireED/requirements.txt
Flask==3.1.2
flask-cors==4.0.0
groq==0.31.1
python-dotenv==1.1.1
pymongo==4.8.0